+++ CHANGELOG.md
@@ -7,9 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-## [0.15.0-a] - 2025-06-10
+## [0.15.0] - 2025-06-10
 ### Added
+- FreeBox type
 - add Polyline.Segment
+- add Scale and ScaleOn methods
+### Changed
+- args for Box.createFromRect2D are curried now
 
 ## [0.14.0] - 2025-05-24
 ### Added